src/embeddings/requirements.txt
fastapi
pydantic-settings
gunicorn
uvicorn
aiohttp
openai
chromadb
asyncpg
langchain-community
langchain-text-splitters
tiktoken
clean-text
unidecode
pandas
numpy
scikit-learn
matplotlib
plotly